requirements.txt
argparse
numpy
openvino
opencv-python
pyqt5
pandas
pyserial
paddlepaddle-gpu
paddlex
paddlehub
requests
rplidar-roboticia
scipy
tqdm
tensorboard
addict
sklearn
pathspec
